% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineMI.R
\name{bind3d}
\alias{bind3d}
\title{Helper function to bind together 3d arrays by the 3d dimension}
\usage{
bind3d(arrayList)
}
\arguments{
\item{arrayList}{list of 3d arrays to bind}
}
\value{
concatenated 3d array
}
\description{
Helper function to bind together 3d arrays by the 3d dimension
}
