% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineMI.R
\name{calcKnots}
\alias{calcKnots}
\title{Calculate B-Spline knots}
\usage{
calcKnots(nBins, splineOrder)
}
\arguments{
\item{nBins}{Number of bins}

\item{splineOrder}{Spline order}
}
\value{
vector of B-Spline knots
}
\description{
Calculate B-Spline knots
}
