% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineMI.R
\name{calcSplineMI}
\alias{calcSplineMI}
\title{Calculate mutual information}
\usage{
calcSplineMI(x, nBins, splineOrder, cores = 1)
}
\arguments{
\item{x}{Gene expression matrix with genes in rows and samples in columns}

\item{nBins}{Number of bins}

\item{splineOrder}{Spline order}

\item{cores}{if >1 then run in parallel}
}
\value{
matrix of mutual information (bits)
}
\description{
Calculate mutual information for an expression matrix using B-Spline binning.
}
\examples{
data( riceEx )
calcSplineMI( riceEx[1:5, ], 7, 3)

# compare single CPU vs two CPU's
system.time( mi_1 <- calcSplineMI( riceEx, 7, 3) )
system.time( mi_2 <- calcSplineMI( riceEx, 7, 3, cores = 2) )
identical( mi_1, mi_2 )
}
