% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineMI.R
\name{calcSplineMItoFile}
\alias{calcSplineMItoFile}
\title{Calculates MI matrix and write directly to file}
\usage{
calcSplineMItoFile(x, nBins, splineOrder, filename, fromRow = 2,
  toRow = nrow(x))
}
\arguments{
\item{x}{Gene expression matrix with genes in rows and samples in columns}

\item{nBins}{Number of bins}

\item{splineOrder}{Spline order}

\item{filename}{name of output file}

\item{fromRow}{first row of MI matrix to generate}

\item{toRow}{last row of MI matrix to generate}
}
\value{
Returns nothing. The results are written to file as text with values
space separated and one row per line. Gene IDs are discarded.
}
\description{
Calculates specified rows of lower half of MI matrix and write directly to
file. This uses less memory as it doesn't need to hold the entire mi matrix
in memory. By setting fromRow and toRow it is possible to split up the job
to run independantly in parallel.
}
\examples{
tmp <- tempfile()
data(riceEx)
calcSplineMItoFile( riceEx, 7, 3, filename = tmp, fromRow = 4, toRow = 6 )
cat(readLines(tmp),sep="\\n")
unlink(tmp)
}
