% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineMI.R
\name{entropy1d}
\alias{entropy1d}
\title{Calculate entropy per gene}
\usage{
entropy1d(weights)
}
\arguments{
\item{weights}{3-dimensional array of weights with dimensions:
nBins x samples x genes. (see \code{\link{calcWeights}})}
}
\value{
vector of entropy per gene in bits
}
\description{
Calculate entropy per gene
}
\examples{
data( riceEx )
hist( entropy1d( calcWeights( riceEx, 7, 3) ) )
}
