% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineMI.R
\name{entropy2d}
\alias{entropy2d}
\title{Calculate joint entropy of two genes}
\usage{
entropy2d(weights, i1, i2)
}
\arguments{
\item{weights}{3-dimensional array of weights with dimensions:
nBins x samples x genes as generated by \code{\link{calcWeights}}}

\item{i1}{index to 1st gene}

\item{i2}{index to 2nd gene}
}
\value{
joint entropy in bits
}
\description{
Calculate joint entropy of two genes
}
\examples{
entropy2d( calcWeights( riceEx, 7, 3), 1, 2 )
}
