% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineMI.R
\name{hist2d}
\alias{hist2d}
\title{Calculate joint probability density matrix from weights}
\usage{
hist2d(weights, i1, i2)
}
\arguments{
\item{weights}{3-dimensional array of weights with dimensions:
nBins x samples x genes as generated by \code{\link{calcWeights}}}

\item{i1}{index to 1st gene}

\item{i2}{index to 2nd gene}
}
\value{
Returns joint probability density matrix [bins x bins]
}
\description{
Calculate joint probability density matrix from weights
}
\examples{
m <- hist2d( calcWeights( riceEx, 7, 3), 1, 2 )
image(m)
}
