% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riceEx}
\alias{riceEx}
\title{Gene expression data rice}
\format{An object of class \code{matrix} with 100 rows and 123 columns.}
\usage{
riceEx
}
\description{
A matrix containing RNAseq gene expression (log2 transformed TPM) from 100
randomly selected highly expressed genes in 123 randomly selected samples.
}
\details{
Genes in rows and samples in columns
}
\keyword{datasets}
