% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineMI.R
\name{scale0to1}
\alias{scale0to1}
\title{Scale a vector to between 0 and 1}
\usage{
scale0to1(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
scaled numeric vector
}
\description{
Scale a vector to between 0 and 1
}
